requirements.txt
asyncio==3.4.3
dateutils==0.6.6
  - argparse [required: Any, installed: 1.1]
  - python-dateutil [required: Any, installed: 2.8.0]
    - six [required: >=1.5, installed: 1.12.0]
  - pytz [required: Any, installed: 2018.9]
discord.py==0.16.12
  - aiohttp [required: >=1.0.0,<1.1.0, installed: 1.0.5]
    - async-timeout [required: Any, installed: 3.0.1]
    - chardet [required: Any, installed: 3.0.4]
    - multidict [required: >=2.0, installed: 4.5.2]
  - websockets [required: >=3.1,<4.0, installed: 3.4]
gspread==3.1.0
  - requests [required: >=2.2.1, installed: 2.21.0]
    - certifi [required: >=2017.4.17, installed: 2019.3.9]
    - chardet [required: >=3.0.2,<3.1.0, installed: 3.0.4]
    - idna [required: >=2.5,<2.9, installed: 2.8]
    - urllib3 [required: >=1.21.1,<1.25, installed: 1.24.1]
oauth2client==4.1.3
  - httplib2 [required: >=0.9.1, installed: 0.12.1]
  - pyasn1 [required: >=0.1.7, installed: 0.4.5]
  - pyasn1-modules [required: >=0.0.5, installed: 0.2.4]
    - pyasn1 [required: >=0.4.1,<0.5.0, installed: 0.4.5]
  - rsa [required: >=3.1.4, installed: 4.0]
    - pyasn1 [required: >=0.1.3, installed: 0.4.5]
  - six [required: >=1.6.1, installed: 1.12.0]
pipdeptree==0.13.2
  - pip [required: >=6.0.0, installed: 18.1]
setuptools==40.6.2